# backend/docs/swagger.yaml
definitions:
  schemas.HTTPError:
    properties:
      message: {}
    type: object
  schemas.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      username:
        type: string
    required:
    - firstname
    - lastname
    - username
    type: object
  schemas.UserToCreate:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - firstname
    - password
    - username
    type: object
  schemas.Users:
    properties:
      users:
        items:
          $ref: '#/definitions/schemas.User'
        type: array
    type: object
host: localhost:1323
info:
  contact:
    email: support@worki.io
    name: Worki Support
  description: This is API for Worki Messanger.
  title: Worki messenger API
  version: "0.1"
paths:
  /users/:
    get:
      consumes:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Users'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.HTTPError'
      summary: Get users
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UserToCreate'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.HTTPError'
      summary: Create user
  /users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.HTTPError'
      summary: Delete user by ID
    get:
      consumes:
      - application/json
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.HTTPError'
      summary: Get user by ID
    put:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.User'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.HTTPError'
      summary: Update user
  /users/{username}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Account Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.HTTPError'
      summary: Delete user by Username
    get:
      consumes:
      - application/json
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.HTTPError'
      summary: Get user by Username
swagger: "2.0"